Implemented SNP handlers
